Converted Report pages(department, group, user) and Dashboard pages(department, group, user, task) to the client components. Remove the redundant force dynamic and revalidatePath from dashboard and report
