+++ src/app/dashboard/user/page.jsx
@@ -1,12 +1,41 @@
-import React from "react";
+"use client";
+
+import React, { useState, useEffect } from "react";
 import UserDashboard from "./UserDashboard";
 import { getAllUser } from "@/model/user";
 import { getAllGroup } from "@/model/group";
-export const dynamic = "force-dynamic";
 
-const User = async ({ searchParams }) => {
-  const users = await getAllUser();
-  const groups = await getAllGroup();
+const User = ({ searchParams }) => {
+  const [users, setUsers] = useState([]);
+  const [groups, setGroups] = useState([]);
+  const [loading, setLoading] = useState(true);
+
+  useEffect(() => {
+    async function fetchData() {
+      try {
+        const [usersData, groupsData] = await Promise.all([
+          getAllUser(),
+          getAllGroup()
+        ]);
+
+        if (usersData && !usersData.error) setUsers(usersData);
+        if (groupsData && !groupsData.error) setGroups(groupsData);
+      } catch (error) {
+        console.error("Error fetching user/group data:", error);
+      } finally {
+        setLoading(false);
+      }
+    }
+    fetchData();
+  }, []);
+
+  if (loading) {
+    return (
+      <div className="min-h-screen flex items-center justify-center">
+        <span className="loading loading-spinner loading-lg text-primary"></span>
+      </div>
+    );
+  }
 
   return (
     <>
